move all logic to index.ts
