+++ dev/src/index.ts
@@ -35,7 +35,7 @@ import {
   validateResourcePath,
 } from './path';
 import {ClientPool} from './pool';
-import {CollectionReference, Query, QueryOptions} from './reference';
+import {ChildDocPath, CollectionReference, Query, QueryOptions} from './reference';
 import {DocumentReference} from './reference';
 import {Serializer} from './serializer';
 import {Timestamp} from './timestamp';
@@ -60,6 +60,7 @@ import {
 import {WriteBatch} from './write-batch';
 
 import api = google.firestore.v1;
+import arrify = require('arrify');
 
 export {
   CollectionReference,
@@ -1528,6 +1529,94 @@ export class Firestore {
     });
 
     return result.promise;
+  }
+
+  async processChildDocs(
+    docs: QueryDocumentSnapshot[] | DocumentSnapshot[], allChilds: string[][]
+  ): Promise<void> {
+    const currentChilds = this.getCurrentChildNames(allChilds);
+    const childDocsPath = this.getChildDocsPath(docs, currentChilds);
+    const distinctDocsPath = [...new Set(childDocsPath.map(doc => doc.path))];
+    const docRef: DocumentReference[] = distinctDocsPath.map(path => {
+      return this.doc(path);
+    });
+    const childDocs = await this.getAll(...docRef);
+    return new Promise(async (resolve, reject) => {
+      try {
+        currentChilds.forEach(child => {
+          docs
+            .filter(cs => cs.protoField(child))
+            .map(doc => {
+              const docsPathToReplace = childDocsPath.filter(docPath => {
+                return docPath.parentId === doc.id && docPath.childEntity === child;
+              });
+              let childToReplace: DocumentSnapshot | DocumentSnapshot[] | undefined;
+              if (docsPathToReplace[0]!.type === 'arrayValue') {
+                childToReplace = childDocs.filter(childDoc => {
+                  return docsPathToReplace.map(path => path.id).includes(childDoc.id);
+                });
+              } else {
+               childToReplace = childDocs.find(childDoc => {
+                  return childDoc.id === docsPathToReplace[0].id;
+                });
+              }
+              doc.setProperty(child, childToReplace!);
+            });
+        });
+        const remainingChild = this.getRemainingChildNames(allChilds);
+        if (remainingChild.length > 0) {
+          await this.processChildDocs(childDocs, remainingChild);
+        }
+        resolve();
+      }
+      catch (e) {
+        reject(e);
+      }
+    });
+  }
+
+  getChildDocsPath(docs: QueryDocumentSnapshot[] | DocumentSnapshot[], child: string[]): ChildDocPath[] {
+    let childDocsPath: ChildDocPath[] = [];
+    child.forEach(childEntity => {
+      const keyValue = docs
+        .filter(cs => cs.protoField(childEntity))
+        .map(doc => {
+          // tslint:disable-next-line: no-any
+          const field = doc.protoField(childEntity) as any;
+          let ref = field;
+          if (field.valueType === 'arrayValue') {
+            ref = field.arrayValue.values;
+          }
+          // tslint:disable-next-line: no-any
+          return (arrify(ref)).map((xs: any) => {
+            const qualifiedResourcePath = QualifiedResourcePath.fromSlashSeparatedString(xs.referenceValue);
+            return {
+              path: qualifiedResourcePath.relativeName,
+              parentId: doc.id,
+              childEntity,
+              type: field.valueType,
+              id: qualifiedResourcePath.id
+            }
+          });
+        });
+      childDocsPath = childDocsPath.concat(...keyValue);
+    });
+    return childDocsPath;
+  }
+
+  getCurrentChildNames(childs: string[][]) {
+    let currentChild= childs.map(xs => {
+      return xs[0];
+    }).filter(x => x.length > 0);
+    currentChild = [...new Set(currentChild)];
+    return currentChild;
+  }
+
+  getRemainingChildNames(childs: string[][]) {
+    return childs.filter(xs => {
+      xs.shift();
+      return xs.length > 0;
+    });
   }
 }
